requirements.txt

grpcio==1.66.1
grpcio-tools==1.66.1
pandas==2.3.3
scikit-learn==1.7.2
joblib
uvloop
